Make flavor of auto-republish files configurable
Assuming which files of a mediapackage are the metadata catalogs and ACLs can
never work perfectly. Therefore, this makes the flavors for selecting configurable,
so users can configure the same flavors they would use in a republish-metadata
workflow.

Although it is not recommended, this change technically allows users to also republish
other files. So it relies on users configuring responsibly.