Add a simple example Spring security filter
I added it to the JWT module,
but that's coincidental.
It can live in any module.

Note that this is not an OSGi service,
but rather a Spring bean.
No `@Reference` to get to the dependencies;
the wiring is done in the security configuration.
Also no `@Activate`,
but I'm sure there is an analogue to that. 🤷‍♀️

The filter just adds **some** role to **some** user reference.
Of course we will have to do something more clever.
Also obviously there are no JWTs involved here, yet.
And maybe we don't even want to abuse user references.
This is really just for demostration purposes.
But it works: This adds said role to said user,
and you can see it reflected in `/info/me.json`
even if that user is authenticated using some other mechanism.
However, this prevents creation of a user with the same name
for example in the admin UI,
so you have to do that before installing this patch. ;D
This is one reason why I think user references are not the right place
to store our roles. :/

Anyway, the config adds the filter to the preauth composite filter.
That's one way to do it, but I don't know if it's idiomatic.
You can also place it in relation to any other filter.
[This](https://docs.spring.io/spring-security/site/docs/4.2.13.RELEASE/reference/htmlsingle/#ns-custom-filters)
was helpful to my understanding. 🤷‍♀️

So what's the trick here?
Why is this called every time while the JWT stuff isn't?
That's kind of having it backwards:
`GenericFilterBean`-s (that are configured correctly) always run on
every request.
(Well, every filter in the chain can abort the chain,
but if you are early enough the above is true.)
It's the magic of the other filter base classes that the JWT module uses
that do some of the heavy lifting of authentication for you,
and also check whether there is already a session.
If so, they rightfully don't do anything.